change proj_id from NUMERIC to SERIAL

+++ ActualProject/sql/createTableSchema.sql
@@ -5,7 +5,7 @@ CREATE TABLE member (
 );
 
 CREATE TABLE project (
-	id NUMERIC(16) PRIMARY KEY CHECK(id >=0),
+	id SERIAL PRIMARY KEY CHECK(id >=0),
 	title VARCHAR(256) NOT NULL,
 	description VARCHAR(2048) DEFAULT '' NOT NULL,
 	category VARCHAR(256) NOT NULL CHECK(category='Fashion' OR category='Technology' OR category='Games' OR category='Food' OR category='Music' OR category='Photography' OR category='Handicraft' OR category='Community'),
@@ -15,7 +15,7 @@ CREATE TABLE project (
 
 CREATE TABLE advertise (
 	entrepreneur VARCHAR(16) REFERENCES member(username) ON DELETE CASCADE,
-	proj_id NUMERIC(16) REFERENCES project(id) ON DELETE CASCADE,
+	proj_id SERIAL REFERENCES project(id) ON DELETE CASCADE,
 	amt_needed NUMERIC(15,2) NOT NULL DEFAULT '0.00' CHECK(amt_needed >= 0),
 	amt_raised NUMERIC(15,2) NOT NULL DEFAULT '0.00' CHECK(amt_raised<=amt_needed AND amt_raised>=0),
 	status INT NOT NULL DEFAULT 0 CHECK(status=0 OR status=1),
@@ -24,7 +24,7 @@ CREATE TABLE advertise (
 
 CREATE TABLE invest (
 	investor VARCHAR(16) REFERENCES member(username) ON DELETE CASCADE,
-	proj_id NUMERIC(16) REFERENCES project(id) ON DELETE CASCADE,
+	proj_id SERIAL REFERENCES project(id) ON DELETE CASCADE,
 	amount NUMERIC(15,2) NOT NULL DEFAULT '0.00' CHECK(amount >= 0),
 	PRIMARY KEY(investor, proj_id)
 );
